[dashboard] Use feature flag
